refactor(Popover): Move state manipulation to hook
